+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Page rank handling
 ### Changed
 - Export to adselect
+- Processing event payments on supply side
+- Computing statistics of publishers
 
 ## [1.3.2] - 2019-11-05
 ### Fixed