+++ CHANGELOG.md
@@ -6,8 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Page rank handling
+### Changed
+- Export to adselect
+- Processing event payments on supply side
+- Computing statistics of publishers
 ### Fixed
 - Processing payment reports while event were not exported
+
+## [1.3.2] - 2019-11-05
+### Fixed
+- Payment details for missing payments
 
 ## [1.3.1] - 2019-10-30
 ### Changed
@@ -251,7 +261,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v1.3.1...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v1.3.2...develop
+[1.3.2]: https://github.com/adshares/adserver/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/adshares/adserver/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/adshares/adserver/compare/v1.2.5...v1.3.0
 [1.2.5]: https://github.com/adshares/adserver/compare/v1.2.4...v1.2.5